requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpath-ng==1.7.0
MarkupSafe==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai==1.98.0
packaging==25.0
pillow==11.3.0
ply==3.11
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-ulid==3.0.0
PyYAML==6.0.2
redis==6.3.0
redisvl==0.8.0
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
